fix: temporary workaround to determine unused OIDC configs
This is to determine if there are any unused OIDC configs and use one if any
exist.  This is because the rosa CLI does not allow creation of a new OIDC
config when any are unused.  This logic will be removed in a later version
so this is simply a temporary workaround until such a time exists.

Signed-off-by: Dustin Scott <dscott@redhat.com>