fix: overlapping variables
The omit function does not work passing into roles the same way it does
into modules.  Need to correct that and use the 'set_fact' task twice.  This
is because it appears Ansible uses the literal variable that is passed in
once for each task, so setting a fact with the same name does not appear to
work as intended.  I also played around with defaulting and that was not
being picked up properly either.

Signed-off-by: Dustin Scott <dscott@redhat.com>